refacore code
